Handle some exception little server refactor
