+++ Server/src/main/java/com/kolesnikov/gamble/manager/session/Session.java
@@ -1,0 +1,7 @@
+package com.kolesnikov.gamble.manager.session;
+
+import java.io.*;
+
+public interface Session {
+    void start(BufferedReader bufferedReader, BufferedWriter bufferedWriter) throws IOException;
+}
