--- Server/src/main/java/com/kolesnikov/gamble/manager/session/SessionExecutor.java
@@ -1,7 +1,0 @@
-package com.kolesnikov.gamble.manager.session;
-
-import java.io.*;
-
-public interface SessionExecutor {
-    void execute(BufferedReader bufferedReader, BufferedWriter bufferedWriter) throws IOException;
-}
