--- Server/src/main/java/com/kolesnikov/gamble/manager/session/SimpleClientSession.java
@@ -1,34 +1,0 @@
-package com.kolesnikov.gamble.manager.session;
-
-import java.io.*;
-import java.net.Socket;
-
-public class SimpleClientSession implements ClientSession {
-    private final Socket userSocket;
-    private final SessionExecutor sessionExecutor;
-
-    public SimpleClientSession(Socket userSocket, SessionExecutor sessionExecutor) {
-        this.userSocket = userSocket;
-        this.sessionExecutor = sessionExecutor;
-    }
-
-    @Override
-    public void run() {
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(userSocket.getInputStream()));
-             BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(userSocket.getOutputStream()))) {
-            sessionExecutor.execute(reader, writer);
-        } catch (
-                IOException e) {
-            System.out.println(e.getMessage());
-        } catch (Exception e) {
-            e.printStackTrace();//todo убрать
-        } finally {
-            try {
-                userSocket.close();
-            } catch (IOException e) {
-                System.out.println(e.getMessage());
-            }
-        }
-
-    }
-}
